move Logger and LogBuffer to own module
