+++ src/logbuffer.js
@@ -1,0 +1,30 @@
+/*
+Copyright 2018 New Vector Ltd
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+module.exports = class LogBuffer {
+    constructor(page, eventName, eventMapper, reduceAsync=false, initialValue = "") {
+        this.buffer = initialValue;
+        page.on(eventName, (arg) => {
+            const result = eventMapper(arg);
+            if (reduceAsync) {
+                result.then((r) => this.buffer += r);
+            }
+            else {
+                this.buffer += result;
+            }
+        });
+    }
+}
